support using this module with bundlers
