+++ example/index.html
@@ -5,7 +5,7 @@
     <title>ssb-validate2-rsjs</title>
   </head>
   <body>
-    <script type="module" src="./main.js"></script>
+    <script type="module" src="./bundle/main.js"></script>
     <h1>ssb-validate2-rsjs</h1>
     <p>Open the console to read success or error messages from WASM code.</p>
   </body>
