+++ karma.conf.js
@@ -7,13 +7,11 @@ module.exports = function (config) {
 
     // frameworks to use
     // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
-    frameworks: ["jasmine", "esm"],
+    frameworks: ["jasmine", "browserify"],
 
     // list of files / patterns to load in the browser
     files: [
-      { pattern: "dist/main.js", type: "module" },
-      { pattern: "comlink.mjs", type: "module" },
-      { pattern: "test/*.js", type: "module" },
+      { pattern: "test/*.js" },
     ],
 
     // list of files / patterns to exclude
@@ -29,7 +27,18 @@ module.exports = function (config) {
 
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
-    preprocessors: {},
+    preprocessors: {
+      'test/**/*.js': ["browserify"]
+    },
+
+    // karma-browserify options: https://www.npmjs.com/package/karma-browserify
+    browserify: {
+      debug: false,
+    },
+
+    proxies: {
+      '/dist': './dist',
+    },
 
     // test results reporter to use
     // possible values: 'dots', 'progress'
