+++ test/data/singleAuthorMsgs.js
@@ -1,4 +1,4 @@
-export default [
+module.exports = [
   {
     value: {
       key: "%txJjB9WAJRQL9DpZkTqq43FPrWY0Yhb2D9TTfkY1GfY=.sha256",
