+++ test/multiAuthorTest.js
@@ -1,23 +1,15 @@
-import * as Comlink from "../comlink.mjs";
-import jsonMsgs from "./multiAuthorMsgs.js";
+const Comlink = require("comlink");
+const jsonMsgs = require("./data/multiAuthorMsgs.js");
+
+// We can't just require ../index because Karma doesn't bundle `new Worker`
+const validate = Comlink.wrap(new Worker("../dist/main.js"));
 
 // map the msg value for each msg in the json array
 let msgs = jsonMsgs.map((m) => m.value);
 
 // test suite for multi-author verification and validation
 describe("test: ", function () {
-  // define validate here so we can access it in all `it` functions
-  var validate;
-
   beforeAll(async function () {
-    // load webworker module
-    const worker = new Worker(new URL("../dist/main.js", import.meta.url), {
-      type: "module",
-    });
-    // load Validator class from worker module
-    const Validator = Comlink.wrap(worker);
-    // instantiate Validator
-    validate = await new Validator();
     // load wasm module and initialize webworkers
     await validate.ready();
   });
